wsh: add support limited search/not in search/vm
One can `(search/not query)` inside a `search/and` arbitrary queries.

E.g:

```
(search/and (search/term "wireditiger") (search/term "database")
            (search/not (search/term "guile")))
```

The query of `search/not` can be whatever valid query, e.g.:

```
(search/and (search/term "wiredtiger") (search/term "database")
            (search/not (search/or (search/term "guile") (search/term "python"))))
```

This is done by computing the result of the query argument of
`search/not` and removing this results from the rest of the
`search/and` clause. I am not sure this is future proof.

Mind the fact that the previous example is equivalent to:

```
(search/and (search/term "wiredtiger") (search/term "database")
            (search/not (search/or (search-term "guile")))
            (search/not (search/or (search-term "python"))))
```

There is also a `(search/make-predicate ctx arg)` which takes a query
as argument. This returns a procedure that can be passed to a filter
over a set of documents. Maybe it's better to implement `search/not`
in terms of `search/make-predicate`. I don't know.